requirements.txt
git+https://github.com/Hrishi2861/pyrofork-2.2.11-peer-fix.git
requests
aria2p
python-dotenv
pytz
tgcrypto
pyrogram
pymongo
psutil
aiohttp
ffmpeg-python  # Python wrapper for FFmpeg
imageio[ffmpeg]  # FFmpeg plugin for imageio (if needed for video processing)
